requirements.txt
Faker
google-cloud-bigquery
google-cloud-pubsub